Move v2 bedspace and premises utilities
