+++ server/controllers/temporary-accommodation/manage/premisesController.ts
@@ -8,7 +8,7 @@ import BedspaceService from '../../../services/bedspaceService'
 import AssessmentsService from '../../../services/assessmentsService'
 import extractCallConfig from '../../../utils/restUtils'
 import { createSubNavArr } from '../../../utils/premisesSearchUtils'
-import { showPropertySubNavArray } from '../../../utils/premisesUtils'
+import { premisesActions, showPropertySubNavArray } from '../../../utils/premisesUtils'
 import {
   InvalidParams,
   catchValidationErrorOrPropogate,
@@ -19,7 +19,6 @@ import {
 } from '../../../utils/validation'
 import { filterProbationRegions } from '../../../utils/userUtils'
 import { parseNumber } from '../../../utils/formUtils'
-import { premisesActions } from '../../../utils/v2/premisesUtils'
 import { DateFormats } from '../../../utils/dateUtils'
 import { addPlaceContext, preservePlaceContext } from '../../../utils/placeUtils'
