+++ server/utils/premisesUtils.ts
@@ -1,25 +1,8 @@
-import type { Cas3Premises } from '@approved-premises/api'
+import { Cas3Premises, Cas3PremisesBedspaceTotals } from '@approved-premises/api'
 import { PageHeadingBarItem, PlaceContext, PremisesShowTabs, type SubNavObj } from '@approved-premises/ui'
 import paths from '../paths/temporary-accommodation/manage'
 import { addPlaceContext } from './placeUtils'
-
-export const premisesActions = (premises: Cas3Premises): Array<PageHeadingBarItem> => {
-  const items = []
-
-  if (premises.status !== 'archived') {
-    items.push({
-      text: 'Add a bedspace',
-      classes: 'govuk-button--secondary',
-      href: paths.premises.bedspaces.new({ premisesId: premises.id }),
-    })
-  }
-
-  if (items.length === 0) {
-    return null
-  }
-
-  return items
-}
+import config from '../config'
 
 export const shortAddress = (premises: Cas3Premises): string => {
   const elements = [premises.addressLine1, premises.town, premises.postcode].filter(element => !!element)
@@ -44,4 +27,67 @@ export const showPropertySubNavArray = (
       active: activeTab === 'bedspaces',
     },
   ]
+}
+
+export const premisesActions = (premises: Cas3Premises, placeContext?: PlaceContext): Array<PageHeadingBarItem> => {
+  const editPremisesPath = paths.premises.edit({ premisesId: premises.id })
+
+  const actions = [
+    {
+      text: 'Edit property details',
+      classes: 'govuk-button--secondary',
+      href: addPlaceContext(editPremisesPath, placeContext),
+    },
+  ]
+
+  if (premises.status === 'online') {
+    const addBedspacePath = paths.premises.bedspaces.new({ premisesId: premises.id })
+
+    actions.push({
+      text: 'Add a bedspace',
+      classes: 'govuk-button--secondary',
+      href: addPlaceContext(addBedspacePath, placeContext),
+    })
+
+    if (premises.endDate) {
+      if (config.flags.cancelScheduledArchiveEnabled) {
+        actions.push({
+          text: 'Cancel scheduled property archive',
+          classes: 'govuk-button--secondary',
+          href: paths.premises.cancelArchive({ premisesId: premises.id }),
+        })
+      }
+    } else {
+      actions.push({
+        text: 'Archive property',
+        classes: 'govuk-button--secondary',
+        href: paths.premises.archive({ premisesId: premises.id }),
+      })
+    }
+  } else if (premises.status === 'archived') {
+    if (premises.scheduleUnarchiveDate) {
+      if (config.flags.cancelScheduledArchiveEnabled) {
+        actions.push({
+          text: 'Cancel scheduled property online date',
+          classes: 'govuk-button--secondary',
+          href: paths.premises.cancelUnarchive({ premisesId: premises.id }),
+        })
+      }
+    } else {
+      actions.push({
+        text: 'Make property online',
+        classes: 'govuk-button--secondary',
+        href: paths.premises.unarchive({ premisesId: premises.id }),
+      })
+    }
+  }
+
+  return actions.sort((a, b) => a.text.localeCompare(b.text))
+}
+
+export function isPremiseScheduledToBeArchived(totals: Cas3PremisesBedspaceTotals): boolean {
+  if (!totals.premisesEndDate) return false
+  const endDate = new Date(totals.premisesEndDate)
+  const now = new Date()
+  return endDate > now && totals.status !== 'archived'
 }
